requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
